Makefile: unify makefiles and kconfig for audio
tegra-alt drivers are intended to be moved to common nvidia
repo. This patch combines the Makefile and Kconfig and also
puts conditional complilation for kernel-4.9, since for audio
the merge can be possible only when kernel-4.4 and kernel-4.9
are unified.

for kernel-4.4, all tegra-alt code is built from nvidia/
for kernel-4.9, t18x specific code is built from nvidia/ and
other drivers are built from kernel-4.9/

Bug 200325738

Change-Id: Id9d1cc9b129f61dc4b52d2e3c5466ec94415517c
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1529135
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>