ASoC: tegra-alt: fix admaif regmap callback func
The regmap read, write and volatile callback functions are not
handling the register address checks correctly. Due to this the
regmap dump from sysfs node was not reflecting the correct state
of the register values. Fix the issue and makesure all regmap
register values are reflecting proper values.

Bug 200500656

Change-Id: Ifb5827d1b86384f835c010138275121a87a397ba
Signed-off-by: Mohan Kumar <mkumard@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2080753
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>