ASoC: tegra-alt: clock cleanup for AHUB drivers
Currently we pass module clocks and related parents from DT bindings. The
driver gets these clock handles and does enable/disable() or sets rate
or parents whenever required. But passing the parent clock is not necessary
always, unless we want to change the rate dynamically.

For example consider the case of i2s device. Possible parents of i2s are,
 * PLLA_OUT0, PLLA1_OUT1, SYNC_CLK, PLLP_OUT0, CLK_M
DT binding currently has entries for few of these clocks. This is not
needed, since we don't have requirement to set the parent dynamically.
Generally it is statically configured. Hence "assgigned-clocks" and
"assigned-clock-parents" DT binding can be used to specify the child and
parent relation.

In summary,
 * devm_clk_get() functions are removed for parent and sync clocks. Only
   for xbar, parent clock would be needed since we change the clock rate
   dynamically depending on playback/capture rate. In most of the cases
   only module clocks are sufficient and are retained.
 * sync clocks are currently not used for L4T/Android. It could be later
   required. Hence DT binding for it will be retained (mostly sync_input
   would be needed). Required driver support will be added later.
 * clk_set_parent() calls are removed, since DT clock binding takes care
   of this.
 * removed clk_ape and clk_apb2ape from xbar driver, since just enable
   or disable of these is performed in the driver, which is not really
   required. Because this is taken care in parent(ACONNECT) device node.
 * Few clock names are updated as below in devm_clk_get()
   * DSPK module clock --> "dspk"
   * I2S module clock  --> "i2s"
   * XBAR parent clock --> "parent"

Bug 200503387

Change-Id: I29b77e32cf1f950d9ea3bb66d27b4a8482bf9926
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2159592
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>