gpu: host1x: Set waiter output reference within lock
With the new fence implementation, we need to assign the reference
within the lock. Otherwise, it can happen that the waiter may get
signaled before the pointer has been set, and the fence implementation
will try to call intr_put_ref with a NULL pointer.

Bug 3237754

Change-Id: I465ec3ecf63c4f179ba8ef9cc5af0892a93f20af
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2474949
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit