nvdla: Fix driver unloading
The NVDLA driver is a missing a call to nvhost_module_deinit() on
removal and so is not releasing resources as expected on removal which
prevents the driver from being unloaded and reloaded correctly. Fix this
by add a call to nvhost_module_deinit() in the driver remove callback.

Bug 3641820

Change-Id: I675bb7b2b31c3194ae30c4162f7f72caf4844243
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2710547
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Amit Sharma (SW-TEGRA) <amisharma@nvidia.com>
Reviewed-by: Arvind M <am@nvidia.com>
Reviewed-by: Praveen K <kpraveen@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
GVS: Gerrit_Virtual_Submit