nvmap: Add empty Makefiles for enabling nvmap build
Add empty makefile and its path for building nvmap.

The source code and makefile with proper entries are
copied from the kernel/nvidia to out folder during build.

This will help to build the module for OOT from the source
code available in the kernel/nvidia which supports multiple
versions of kernel and avoid duplication of code.

Bug 3583631

Change-Id: I4ded3c1e49fc60e58d205a0b5e646e7bb616d8c8
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2712957
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
GVS: Gerrit_Virtual_Submit