spi: aurix: get rid of usage of do_exit()
The function do_exit() is exiting from the thread so
it is more appropriate to use the function complete_and_exit().
This function is also chnaged to more appropriate name as
kthread_complete_and_exit() for the K5.17 and beyond.

Use appropriate function name based on the kernel version.

Bug 3604749

Change-Id: I3e72d99a414317b782da2170163519eb2f1cfff3
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2730029
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>