device-tree/Makefile: Tune relative path
The relative path from the device-tree/platform to $TOP
varies between local and GVS build. The folder "nvidia"
in out folder is introduced in the local build which is
not there in the GVs build.

Check the relative path of $TOP and adjust accordingly.

Bug 3787936

Change-Id: I5a34e64a8903f662153c361c2fc7e3ee432d6c14
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2775081
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>