nvidia: oot: Simplify source copying
The script for building OOT modules has been updated to use rsync to
copy the necessary files for drivers outside of the linux-nv-oot
repository. This simplifies the script that copies the source and now
the update the kernel-src-files-copy-list.txt so that it simply contains
a list of source files and directories to copy.

Bug 3820317

Change-Id: I6295ba3c1f074a5e047cb5e00ca41db5366d411b
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2787176
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>