ASoC: tegra: Fix uninitialized array index read
The index of the override arrays is related to stream direction.
The variable used to read these array index is not initialized
and is reported by coverity scan. The variable assignment code
is misplaced. Fix this by moving this code to make sure that
the array index read happens after the variable assignment.

CID 10162269
Bug 3952896

Change-Id: I7e8d000860e799a93437d6eb9d1b788cda118c14
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2860609
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>