mfd: nvvrs-pseq: Fix automatic module loading
The NVIDIA VRS PSEQ driver is not automatically loaded on boot and this
is because the MODULE_DEVICE_TABLE is missing in the driver. Add the
MODULE_DEVICE_TABLE for this driver so that the driver is loaded as
expected on boot.

Bug 3583627

Change-Id: I19316ac7fb26e9cced65d0d7504a5c7f3f49b35f
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2865056
(cherry picked from commit fe3aea29adff6f548dfd29421aa8fa1811401daf)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2866991
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>