pva: Remove dependency on CVNAS
Tegra234 does not support CVSRAM and therefore does not need the CVNAS
driver. When building the PVA driver as an out-of-tree module, we only
need to support Tegra234 and so we can compile out support for CVNAS and
remove the dependency on CVNAS. Note that CONFIG_TEGRA_CVNAS is not set
when building out-of-tree modules.

Bug 4037930

Change-Id: I2f4a524d50229c4a3b42c4b62df67b8a40daeca6
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2903055
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: Omar Nemri <onemri@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>
Tested-by: Omar Nemri <onemri@nvidia.com>