platform: tegra: rtcpu: Don't register firmware
The Tegra firmware-class driver is being deprecated to reduce the number
of out-of-tree modules that are required. The firmware is registered
with this driver for informational purposes. The firmware version is
already printed on boot and so remove the call to register the firmware.

Bug 4090910

Change-Id: I69045a641f6daf21f15f71ae2ea70def2c74b28e
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2897319
Reviewed-by: Matti Ryttylainen <mryttylainen@nvidia.com>
Reviewed-by: Ankur Pawar <ankurp@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2898617
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>