staging: gte_test: Add missing header
With Linux v6.4 the gte_test driver fails to build and the following
error is observed ...

 drivers/staging/platform/tegra/gte_test/tegra194_gte_test.c:67:14:
 error: implicit declaration of function ‘of_find_compatible_node’
 [-Werror=implicit-function-declaration]
   np = of_find_compatible_node(NULL, NULL, "nvidia,tegra194-gte-aon");
        ^~~~~~~~~~~~~~~~~~~~~~~

The function of_find_compatible_node() is not found because the of.h
header is not explicitly included. Fix the build for Linux v6.4 by
including of.h.

Bug 4183168

Change-Id: I5dc6bb9aeb4e20a0bdab7e36c8ef0e5e466ad7ff
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2979248
Reviewed-by: Dipen Patel <dipenp@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>