ASoC: tegra: Only compile drivers if ASoC is enabled
When compiling the out-of-tree audio drivers against a kernel that does
not have CONFIG_SND_SOC enabled, the build fails because the various
snd_soc_xxx functions are not exported by the kernel. Update the
out-of-tree Makefile to only build the Tegra ASoC drivers if
CONFIG_SND_SOC is enabled. Note that CONFIG_SOUND is enabled if
CONFIG_SND_SOC is enabled so it is not necessary to check for both.

Bug 4221847

Change-Id: I7a55e18324baa229d715fa0831df919502e2a3d3
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2993768
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>