nvidia-oot: Fix ccflag for IVC legacy support
The flag CONFIG_TEGRA_IVC_LEGACY_DISABLE was added so that it can be set
by an external Makefile that build the out-of-tree drivers. If this flag
is set by an external Makefile then the ccflags definition
'-DCONFIG_TEGRA_IVC_LEGACY_DISABLE' does not get set as expected and
builds fail. Fix this by setting the ccflag if the Make flag
CONFIG_TEGRA_IVC_LEGACY_DISABLE is set and not just for Linux v6.2+
kernels.

Bug 4221847

Change-Id: Ia662a9e6bfac064fa0571af7f0173cda2cf94b02
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3005233
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>